+++ __external__/services/src/__tests__/User.service.spec.ts
@@ -15,19 +15,12 @@ import {
   UserRole,
   Innovation,
   Role,
-  TouType,
-  TermsOfUseUser,
-  TermsOfUse,
 } from "@domain/index";
 import { InvalidParamsError, InvalidUserTypeError } from "@services/errors";
 import * as dotenv from "dotenv";
 import * as path from "path";
 import { getConnection, getRepository } from "typeorm";
-import {
-  closeTestsConnection,
-  setupTestsConnection,
-  TermsOfUseService,
-} from "..";
+import { closeTestsConnection, setupTestsConnection } from "..";
 import * as helpers from "../helpers";
 import { ProfileModel } from "../models/ProfileModel";
 import { AccessorService } from "../services/Accessor.service";
@@ -50,21 +43,27 @@ describe("User Service Suite", () => {
   let organisationService: OrganisationService;
   let organisation: Organisation;
   let organisationUnit: OrganisationUnit;
-  let touService: TermsOfUseService;
 
   beforeAll(async () => {
-    await setupTestsConnection();
+    //await setupTestsConnection();
     dotenv.config({
       path: path.resolve(__dirname, "./.environment"),
     });
     userService = new UserService(process.env.DB_TESTS_NAME);
     accessorService = new AccessorService(process.env.DB_TESTS_NAME);
     organisationService = new OrganisationService(process.env.DB_TESTS_NAME);
-    touService = new TermsOfUseService(process.env.DB_TESTS_NAME);
 
     organisation = await fixtures.createOrganisation(OrganisationType.ACCESSOR);
     organisationUnit = await fixtures.createOrganisationUnit(organisation);
     await fixtures.createAdminRole();
+
+    jest
+      .spyOn(helpers, "authenticateWitGraphAPI")
+      .mockResolvedValue(":access_token");
+    jest.spyOn(helpers, "getUserFromB2CByEmail").mockResolvedValue({
+      id: "C7095D87-C3DF-46F6-A503-001B083F4630",
+      displayName: ":userName",
+    });
   });
 
   afterAll(async () => {
@@ -74,7 +73,7 @@ describe("User Service Suite", () => {
 
     await query.from(Role).execute();
 
-    closeTestsConnection();
+    //closeTestsConnection();
   });
 
   afterEach(async () => {
@@ -88,11 +87,7 @@ describe("User Service Suite", () => {
     await query.from(OrganisationUser).execute();
     await query.from(Innovation).execute();
     await query.from(UserRole).execute();
-    await query.from(TermsOfUseUser).execute();
     await query.from(User).execute();
-    await query.from(TermsOfUse).execute();
-
-    jest.resetAllMocks();
   });
 
   it("should instantiate the User service", async () => {
@@ -187,108 +182,6 @@ describe("User Service Suite", () => {
     expect(actual.displayName).toEqual("Accessor A");
     expect(actual.type).toEqual(UserType.ACCESSOR);
     expect(actual.organisations.length).toBeGreaterThan(0);
-  });
-
-  it("should retrieve a user profile with terms of use accepted", async () => {
-    // Arrange
-    jest
-      .spyOn(helpers, "authenticateWitGraphAPI")
-      .mockResolvedValue(":access_token");
-
-    jest.spyOn(helpers, "getUserFromB2CByEmail").mockResolvedValue({
-      id: "C7095D87-C3DF-46F6-A503-001B083F4630",
-      displayName: ":userName",
-    });
-
-    jest.spyOn(helpers, "getUserFromB2C").mockResolvedValue({
-      displayName: "Innovator A",
-      identities: [
-        {
-          signInType: "emailAddress",
-          issuerAssignedId: "test_user@example.com",
-        },
-      ],
-      mobilePhone: "+351960000000",
-    });
-
-    let actual: ProfileModel;
-    let err;
-
-    const innovatorUser = await fixtures.createInnovatorUser();
-
-    const innovatorRequestUser = fixtures.getRequestUser(innovatorUser);
-
-    const newToU = await touService.createTermsOfUse(dummy.requestUser, {
-      name: "TERMS OF USE",
-      summary: "TEST NAME",
-      touType: TouType.INNOVATOR,
-      releasedAt: new Date(),
-    });
-
-    try {
-      await touService.acceptTermsOfUse(innovatorRequestUser, newToU.id);
-    } catch (error) {
-      err = error;
-    }
-
-    // Act
-    try {
-      actual = await userService.getProfile(
-        innovatorRequestUser.id,
-        innovatorRequestUser.externalId
-      );
-    } catch (error) {
-      err = error;
-    }
-
-    // Assert
-    expect(err).toBeUndefined();
-    expect(actual.isTouAccepted).toEqual(true);
-  });
-
-  it("should retrieve a user profile with terms of use not accepted yet", async () => {
-    // Arrange
-    jest
-      .spyOn(helpers, "authenticateWitGraphAPI")
-      .mockResolvedValue("access_token");
-    jest.spyOn(helpers, "getUserFromB2C").mockResolvedValue({
-      displayName: "Innovator A",
-      identities: [
-        {
-          signInType: "emailAddress",
-          issuerAssignedId: "test_user@example.com",
-        },
-      ],
-      mobilePhone: "+351960000000",
-    });
-
-    let actual: ProfileModel;
-    let err;
-
-    const innovatorUser = await fixtures.createInnovatorUser();
-
-    const innovatorRequestUser = fixtures.getRequestUser(innovatorUser);
-
-    const newToU = await touService.createTermsOfUse(dummy.requestUser, {
-      name: "TERMS OF USE",
-      summary: "TEST NAME",
-      touType: TouType.INNOVATOR,
-      releasedAt: new Date(),
-    });
-
-    // Act
-    try {
-      actual = await userService.getProfile(
-        innovatorRequestUser.id,
-        innovatorRequestUser.externalId
-      );
-    } catch (error) {
-      err = error;
-    }
-
-    // Assert
-    expect(err).toBeUndefined();
-    expect(actual.isTouAccepted).toEqual(false);
   });
 
   it("should retrieve a user B2C profile information", async () => {
@@ -422,11 +315,6 @@ describe("User Service Suite", () => {
   });
 
   it("should create an accessor user", async () => {
-    jest.spyOn(helpers, "getUserFromB2CByEmail").mockResolvedValue({
-      id: "C7095D87-C3DF-46F6-A503-001B083F4630",
-      displayName: ":userName",
-    });
-
     const result = await userService.createUser(dummy.requestUser, {
       type: UserType.ACCESSOR,
       name: ":name",
@@ -442,11 +330,6 @@ describe("User Service Suite", () => {
   });
 
   it("should create an assessment user", async () => {
-    jest.spyOn(helpers, "getUserFromB2CByEmail").mockResolvedValue({
-      id: "C7095D87-C3DF-46F6-A503-001B083F4630",
-      displayName: ":userName",
-    });
-
     const result = await userService.createUser(dummy.requestUser, {
       type: UserType.ASSESSMENT,
       name: ":name",
@@ -900,10 +783,6 @@ describe("User Service Suite", () => {
 
   it("should update user's organisation unit", async () => {
     // Arrange
-    jest
-      .spyOn(helpers, "authenticateWitGraphAPI")
-      .mockResolvedValue("access_token");
-
     const user = await fixtures.createAccessorUser();
 
     const organisation = await fixtures.createOrganisation(
@@ -944,10 +823,6 @@ describe("User Service Suite", () => {
 
   it("It should get User by external id", async () => {
     // Arrange
-    jest
-      .spyOn(helpers, "authenticateWitGraphAPI")
-      .mockResolvedValue("access_token");
-
     const dbUser = await fixtures.createInnovatorUser();
     // Act
     let err;
